Moved /components to /lib
