+++ src/lib/components/Counter.svelte
